requirements.txt
GitPython==3.1.12
jinja2=2.11.3